% Auto-generated: do not edit by hand
\name{dashDatetimepicker}

\alias{dashDatetimepicker}

\title{DashDatetimepicker component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashDatetimepicker(id=NULL, value=NULL, startDate=NULL, endDate=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{value}{Lists containing elements 'startdate', 'enddate'.
those elements have the following types:
  - startdate (character; optional)
  - enddate (character; optional). The value displayed in the input.}

\item{startDate}{Character. The value displayed in the input.}

\item{endDate}{Character. The value displayed in the input.}
}
