% Auto-generated: do not edit by hand
\name{dashDatetimepicker}

\alias{dashDatetimepicker}

\title{DashDatetimepicker component}

\description{
Dash Datetime Range Picker is a component for selecting two dates (including times). It takes `startDate` and `endDate` as main properties and renders two inputs which open datepickers when they are clicked.
}

\usage{
dashDatetimepicker(id=NULL, endDate=NULL, startDate=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{endDate}{Character. The endDate of the range picker. It will fire a dash callback if it is updated.}

\item{startDate}{Character. The startDate of the range picker. It will fire a dash callback if it is updated.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

